AG-13110 Move calcPanToBBox to src/util/
